requirements.txt
twitchio
python-dotenv
requests
beautifulsoup4